+++ remote/python_vision/requirements.txt
