+++ remote/src/main/java/vision/helperClasses/ContourSet.java
@@ -1,4 +1,4 @@
-package org.example.helperClasses;
+package vision.helperClasses;
 
 import org.opencv.core.MatOfPoint;
